#LANG-1145 set improvement implementation to default
Remove old ArchUtils and rename ArchUtilsImproved to ArchUtils.
